+++ resources/js/components/utils/Notifications.vue
@@ -1,7 +1,15 @@
 <template>
-  <div class="has-margin-3 cursor-pointer" @click="open=true" v-if="countCards()">
-    <i class="fad fa-bell" style="font-size: 1.25em;"></i>
-    <span class="tag is-danger" style="font-size: 0.7em;" v-html="countCards()"></span>
+  <div
+    class="has-margin-3 cursor-pointer"
+    @click="open = true"
+    v-if="countCards()"
+  >
+    <i class="fad fa-bell" style="font-size: 1.25em"></i>
+    <span
+      class="tag is-danger"
+      style="font-size: 0.7em"
+      v-html="countCards()"
+    ></span>
 
     <b-sidebar
       type="is-light"
@@ -13,14 +21,22 @@
       icon-pack="fa"
       mobile="fullwidth"
     >
-      <div class="close-button" @click="open=false">
+      <div class="close-button" @click="open = false">
         <button class="button is-dark">x</button>
       </div>
       <div>
         <div v-for="(line, index) in pending" :key="index">
-          <div class="columns has-all-centered" v-for="(card, indexC) in line.cards" :key="indexC">
+          <div
+            class="columns has-all-centered"
+            v-for="(card, indexC) in line.cards"
+            :key="indexC"
+          >
             <div class="column is-narrow">
-              <show-card style="zoom: .75" :card="card" admin="false"></show-card>
+              <show-card
+                style="zoom: 0.75"
+                :card="card"
+                admin="false"
+              ></show-card>
             </div>
             <div class="column is-narrow is-flex has-all-centered">
               <h4 class="is-size-4">{{ line.student.name }}</h4>
@@ -28,13 +44,26 @@
             <div class="column is-flex is-narrow has-all-centered">
               <button
                 class="button is-success"
-                @click="setCard(card.id, line, card.pivot.marked, true, index, indexC)"
+                @click="
+                  setCard(card.id, line, card.pivot.marked, true, index, indexC)
+                "
                 v-text="getText(card.pivot.marked)"
               ></button>
               <button
                 class="button is-danger has-margin-left-2"
-                @click="setCard(card.id, line, card.pivot.marked, false, index, indexC)"
-              >Cancel</button>
+                @click="
+                  setCard(
+                    card.id,
+                    line,
+                    card.pivot.marked,
+                    false,
+                    index,
+                    indexC
+                  )
+                "
+              >
+                Cancel
+              </button>
             </div>
           </div>
         </div>
@@ -45,13 +74,19 @@
 <script>
 export default {
   props: ["pending"],
-  mounted() {},
-  data: function() {
+  mounted() {
+    this.cards = this.pending;
+    for (let i = 0; i < this.cards.length; i++) {
+      this.cards[i].cards = Object.values(this.cards[i].cards);
+    }
+  },
+  data: function () {
     return {
       open: false,
       overlay: true,
       fullheight: true,
-      fullwidth: false
+      fullwidth: false,
+      cards: [],
     };
   },
   methods: {
@@ -63,22 +98,24 @@ export default {
       }
     },
     setCard(id, line, type, action, index, indexC) {
-      this.pending[index].cards.splice(indexC, 1);
-      if (!this.pending[index].cards.length) this.open = false;
+      if (!action) {
+        this.cards[index].cards.splice(indexC, 1);
+        if (!this.cards[index].cards.length) this.open = false;
+      }
       this.$forceUpdate();
       axios
         .post("/classroom/card/usedelete/" + id, {
           student: line.student.id,
           action: action,
-          type: type
+          type: type,
         })
-        .then(response => {
+        .then((response) => {
           this.$toasted.show(response.data.message, {
             position: "top-center",
             duration: 3000,
             iconPack: "fontawesome",
             icon: response.data.icon,
-            type: response.data.type
+            type: response.data.type,
           });
           if (action) {
             if (response.data.type == "success") location.reload();
@@ -87,12 +124,12 @@ export default {
     },
     countCards() {
       let count = 0;
-      this.pending.forEach(element => {
+      this.pending.forEach((element) => {
         count += _.size(element.cards);
       });
       return count;
-    }
+    },
   },
-  computed: {}
+  computed: {},
 };
 </script>